+++ pkg/bgp/mp-reach-nlri.go
@@ -13,6 +13,7 @@ import (
 	"github.com/sbezverk/gobmp/pkg/l3vpn"
 	"github.com/sbezverk/gobmp/pkg/ls"
 	"github.com/sbezverk/gobmp/pkg/mcastvpn"
+	"github.com/sbezverk/gobmp/pkg/rtc"
 	"github.com/sbezverk/gobmp/pkg/srpolicy"
 	"github.com/sbezverk/gobmp/pkg/unicast"
 	"github.com/sbezverk/tools"
@@ -189,6 +190,15 @@ func (mp *MPReachNLRI) GetFlowspecNLRI() (*flowspec.NLRI, error) {
 func (mp *MPReachNLRI) GetNLRIMCASTVPN() (*mcastvpn.Route, error) {
 	if mp.SubAddressFamilyID == 5 {
 		return mcastvpn.UnmarshalMCASTVPNNLRI(mp.NLRI)
+	}
+
+	return nil, fmt.Errorf("not found")
+}
+
+// GetNLRIRTC checks for presence of NLRI 132 Route Target Constraint in the NLRI 14 NLRI data and if exists, instantiate RTC NLRI object
+func (mp *MPReachNLRI) GetNLRIRTC() (*rtc.Route, error) {
+	if mp.SubAddressFamilyID == 132 {
+		return rtc.UnmarshalRTCNLRI(mp.NLRI)
 	}
 
 	return nil, fmt.Errorf("not found")